add missing tests
